Add scale on hover to all base button icons
